fixes in serial protocol
- fixed retrieval of tokens from in put buffer (probably needs a
complete refactoring, the current scheme is weird)
- this enabled complete removal of delays between bytes and words on
the msx side
- added createFile, renameFile, deleteFile functions
- stubs for sequentialWrite and absSectorRead (probably the latter is
never to be implemented)

+++ sw/spyvspy/dosglob.h
@@ -57,7 +57,7 @@ public:
         if (Util::dosname(fname, dosname)) {
           //info("["); for (int q = 0; q < 11; q++) info("%c", m_pattern[q]); info("]");
           for (i = 0; 
-                i < sizeof(dosname) &&
+                i < (int)sizeof(dosname) &&
                 (m_pattern[i] == dosname[i] || m_pattern[i] == '?'); 
                 i++);
         }
